solved the possible unhandled promise rejection
when I moved the aggregatefilteredrecords function out of dashboard and into functions, I had to then pass the setFilteredItems function in as a callback.  However I never updated the functions in datatable to add the new required arumenent so they were calling aggregatedfilteredrecords with insufficient arguments.  the function would have been trying to execute a function that it didnt recognize